refactor: skills controllers refactored
