+++ backend/src/controllers/skills/index.ts
@@ -1,0 +1,1 @@
+export { default as getAllSkills } from './getAllSkills/getAllSkills';
